컴포넌트로 묶여 있던 모달과 버튼 분리, 모달과 버튼을 묶어두었던 컴포넌트는 삭제
